requirement.txt
aniso8601==7.0.0
APScheduler==3.6.1
Click==7.0
dsnparse==0.1.12
elasticsearch==7.0.2
elasticsearch-dsl==7.0.0
Flask==0.12.2
Flask-APScheduler==1.11.0
Flask-Login==0.4.1
Flask-RESTful==0.3.7
force-db==0.1
funcsigs==1.0.2
futures==3.3.0
gunicorn==19.7.1
honcho==1.0.1
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
PyMySQL==0.6.7
python-dateutil==2.8.0
python-envcfg==0.2.0
pytz==2019.1
six==1.12.0
SQLAlchemy==1.0.11
tzlocal==2.0.0
urllib3==1.25.3
Werkzeug==0.15.5
